TODO: fix SearchStoreItem function - seems to not pass in the vector
ADD:
the story - start in a tavern where you are given a quest to enter the dungeon
first start by going to the store to buy supplies then go into the dungeon. when you return you can buy other items and upgrades to go farther in the dungeon until you get to the boss and beat the dungeon.